Organizes and updates tests for connectors
