+++ tests/test_connectors/test_input.py
@@ -1,67 +1,10 @@
-"""Tests the connectivity and functionality of ``Input`` connector objects."""
+"""Tests the functionality of ``Input`` connector objects."""
 
 import time
 from unittest import TestCase
 
-from egon.connectors import Input, KillSignal, Output
+from egon.connectors import Input, KillSignal
 from egon.mock import MockSource, MockTarget
-
-
-class InstanceConnections(TestCase):
-    """Test the connection of generic connector objects to other"""
-
-    def setUp(self) -> None:
-        """Define a connector instances for testing"""
-
-        self.input_connector = Input()
-        self.output_connector = Output()
-
-    def test_input_is_connected_boolean(self) -> None:
-        """The ``is_connected`` attribute returns the current connection state"""
-
-        self.assertFalse(self.input_connector.is_connected)
-        self.output_connector.connect(self.input_connector)
-        self.assertTrue(self.input_connector.is_connected)
-
-    def test_multiple_output_support(self):
-        """Test inputs support the accumulation of data from multiple outputs"""
-
-        # Create two nodes to output data into a single receiving node
-        test_data = [1, 2, 3, 4, 5, 6]
-        source_1 = MockSource(test_data[:3])
-        source_2 = MockSource(test_data[3:])
-        target = MockTarget()
-
-        # Connect the two outputs to feed the same input
-        source_1.output.connect(target.input)
-        source_2.output.connect(target.input)
-
-        # Execute the nodes so data is passed through the connectors
-        source_1.execute()
-        source_2.execute()
-        target.execute()
-
-        self.assertCountEqual(test_data, target.accumulated_data)
-
-
-class PartnerMapping(TestCase):
-    """Test connectors with an established connection correctly map to neighboring connectors/nodes"""
-
-    def setUp(self) -> None:
-        """Create two connected pipeline elements"""
-
-        self.input = Input()
-        self.output1 = Output()
-        self.output2 = Output()
-
-        self.output1.connect(self.input)
-        self.output2.connect(self.input)
-
-    def test_is_aware_of_partners(self) -> None:
-        """Test connectors map to the correct partner connector"""
-
-        output_connectors = [self.output1, self.output2]
-        self.assertCountEqual(output_connectors, self.input.get_partners())
 
 
 class InputGet(TestCase):
@@ -108,9 +51,6 @@ class InputIterGet(TestCase):
     """Test iteration behavior of the ``iter_get`` method"""
 
     def setUp(self) -> None:
-        """Define a node with an attached ``Input`` instance"""
-
-        # Create a node with an input connector
         self.target = MockTarget()
 
     def test_raises_stop_iteration_on_kill_signal(self) -> None:
@@ -164,8 +104,6 @@ class QueueProperties(TestCase):
     """Test  test the exposure of queue properties by the overlying ``Connector`` class"""
 
     def setUp(self) -> None:
-        """Create a ``DataStore`` instance"""
-
         self.connector = Input(maxsize=1)
 
     def test_size_matches_queue_size(self) -> None:
@@ -192,3 +130,27 @@ class QueueProperties(TestCase):
         time.sleep(1)
 
         self.assertFalse(self.connector.empty())
+
+
+class InputInstanceConnections(TestCase):
+    """Test the connection of generic connector objects to other"""
+
+    def test_multiple_output_support(self):
+        """Test inputs support the accumulation of data from multiple outputs"""
+
+        # Create two nodes to output data into a single receiving node
+        test_data = [1, 2, 3, 4, 5, 6]
+        source_1 = MockSource(test_data[:3])
+        source_2 = MockSource(test_data[3:])
+        target = MockTarget()
+
+        # Connect the two outputs to feed the same input
+        source_1.output.connect(target.input)
+        source_2.output.connect(target.input)
+
+        # Execute the nodes so data is passed through the connectors
+        source_1.execute()
+        source_2.execute()
+        target.execute()
+
+        self.assertCountEqual(test_data, target.accumulated_data)
